+++ recipe/dashboard/dashboardmodels/models.go
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
+/* Copyright (c) 2022, VRAI Labs and/or its affiliates. All rights reserved.
  *
  * This software is licensed under the Apache License, Version 2.0 (the
  * "License") as published by the Apache Software Foundation.
@@ -13,18 +13,19 @@
  * under the License.
  */
 
-package supertokens
+package dashboardmodels
 
-const (
-	HeaderRID = "rid"
-	HeaderFDI = "fdi-version"
-)
+type TypeInput struct {
+	ApiKey   string
+	Override *OverrideStruct
+}
 
-// VERSION current version of the lib
-const VERSION = "0.9.0"
+type TypeNormalisedInput struct {
+	ApiKey   string
+	Override OverrideStruct
+}
 
-var (
-	cdiSupported = []string{"2.8", "2.9", "2.10", "2.11", "2.12", "2.13", "2.14", "2.15"}
-)
-
-const DashboardVersion = "0.1"
+type OverrideStruct struct {
+	Functions func(originalImplementation RecipeInterface) RecipeInterface
+	APIs      func(originalImplementation APIInterface) APIInterface
+}