+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Adds:
+
+- Adds Dashboard recipe
+
+
 ## [0.9.0] - 2022-09-14
 
 ### Added